test: split chatService.test.ts into concern-scoped files
Break the 99K-line chatService.test.ts into 7 focused files:
conversations, messages, workspace, migration, usage, workspaceMemory,
and helpers. Each split reuses the workspaceHash helper instead of
redefining it inline. All 190 tests still pass.
